Make DeviceEditFragment and DeviceUpdateAvailableFragment wider
